+++ source/ed/build_helper.h
@@ -56,11 +56,11 @@ struct VJ {
     /// Ajout un nouveau stopTime
     /// Lorsque le depart n'est pas specifié, on suppose que c'est le même qu'à l'arrivée
     /// Si le stopPoint n'est pas connu, on le crée avec un stopArea ayant le même nom
-    VJ& operator()(const std::string &stopPoint, int arrivee, int depart = -1, uint32_t local_traffic_zone = std::numeric_limits<uint32_t>::max(),
+    VJ& operator()(const std::string &stopPoint, int arrivee, int depart = -1, uint16_t local_traffic_zone = std::numeric_limits<uint16_t>::max(),
                    bool drop_off_allowed = true, bool pick_up_allowed = true);
 
     VJ& operator()(const std::string &stopPoint, const std::string& arrivee, const std::string& depart,
-            uint32_t local_traffic_zone = std::numeric_limits<uint32_t>::max(), bool drop_off_allowed = true, bool pick_up_allowed = true);
+            uint16_t local_traffic_zone = std::numeric_limits<uint16_t>::max(), bool drop_off_allowed = true, bool pick_up_allowed = true);
 
     /// Transforme les horaires en horaire entre start_time et end_time, toutes les headways secs
     VJ & frequency(uint32_t start_time, uint32_t end_time, uint32_t headway_secs);
